Major:
Added ARMA Generation

Added invert check and minor root function.

Modified:
ARMA Guesses now use the invert_check method

Minor:

Added comments and changed summing properties of sum_field_vec


Former-commit-id: ced77874f8217fba9eea5fa9cd2d2c6ce2c03c90

+++ src/armadillo_manipulations.h
@@ -9,6 +9,6 @@ arma::vec reverse_vec(arma::vec x);
 
 arma::mat field_to_matrix(arma::field<arma::vec> x);
 
-unsigned int sum_field_vec(const arma::field<arma::vec>& x);
+double sum_field_vec(const arma::field<arma::vec>& x);
 
 #endif
